backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0

# Image Processing Libraries
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0
matplotlib==3.7.2

# Additional utilities
aiofiles==23.2.1
uuid==1.30
